+++ packages/integration-sdk-testing/src/index.ts
@@ -2,3 +2,4 @@ export * from './context';
 export * from './logger';
 export * from './recording';
 export * from './jobState';
+export * from './jest';