+++ packages/integration-sdk/CHANGELOG.md
@@ -14,6 +14,48 @@ and this project adheres to
 - Fixed `visualize` cmd where mapped relationships did not consider `targetFilterKey` when matching entities
 - Fixed `visualize` cmd where multiple nodes with the same nodeId could be created, which causes rendering to fail
 
+## 2.8.0 - 2020-08-3
+
+### Added
+
+- Automatically register custom Jest test matchers when using
+  `@jupiterone/integration-sdk-dev-tools`. See:
+  https://github.com/JupiterOne/sdk/issues/265
+
+## 2.7.0 - 2020-08-2
+
+### Added
+
+- Added Jest test matcher for validating collected entities against a JSON
+  schema. See: https://github.com/JupiterOne/sdk/issues/263
+
+Example:
+
+```typescript
+expect(context.jobState.collectedEntities).toMatchGraphObjectSchema({
+  _class: ['Service'],
+  schema: {
+    additionalProperties: false,
+    properties: {
+      _type: { const: 'google_cloud_api_service' },
+      category: { const: ['infrastructure'] },
+      state: {
+        type: 'string',
+        enum: ['STATE_UNSPECIFIED', 'DISABLED', 'ENABLED'],
+      },
+      enabled: { type: 'boolean' },
+      usageRequirements: {
+        type: 'array',
+        items: { type: 'string' },
+      },
+      _rawData: {
+        type: 'array',
+        items: { type: 'object' },
+      }
+   }
+});
+```
+
 ## 2.6.0 - 2020-07-28
 
 ### Added
@@ -23,7 +65,8 @@ and this project adheres to
 
 ### Changed
 
-- Deprecated `createIntegrationRelationship`. Developers should use the exported `createDirectRelationship` or `createMappedRelationship` functions.
+- Deprecated `createIntegrationRelationship`. Developers should use the exported
+  `createDirectRelationship` or `createMappedRelationship` functions.
 
 ## 2.5.0 - 2020-07-28